Rearrange things to wait for MySQL in initial row stream and copy phase cycles
